Put fetch errors in fetch trace and standardize
