+++ frontend/src/hooks.client.ts
@@ -62,11 +62,15 @@ function shouldTryAutoReload(updateDetected: boolean): boolean {
  * This is obviously NOT a SvelteKit handler/feature. It just mimics the `handleFetch` in hooks.server.ts.
  */
 handleFetch(async ({ fetch, args }) => {
-  const response = await traceFetch(() => fetch(...args));
+  const response = await traceFetch(async () => {
+    const response = await fetch(...args);
 
-  validateFetchResponse(response,
-    location.pathname === '/login',
-    location.pathname === '/' || location.pathname === '/home' || location.pathname === '/admin');
+    validateFetchResponse(response,
+      location.pathname === '/login',
+      location.pathname === '/' || location.pathname === '/home' || location.pathname === '/admin');
+
+    return response;
+  });
 
   if (response.headers.get('lexbox-refresh-jwt') == 'true') {
     await invalidate(USER_LOAD_KEY);
