+++ frontend/src/lib/otel/otel.client.ts
@@ -1,37 +1,16 @@
 import { BatchSpanProcessor, WebTracerProvider } from '@opentelemetry/sdk-trace-web'
-import { SERVICE_NAME, ensureErrorIsTraced, tracer } from '.'
 
 import { APP_VERSION } from '$lib/util/version';
 import { OTLPTraceExporterBrowserWithXhrRetry } from './trace-exporter-browser-with-xhr-retry';
 import { Resource } from '@opentelemetry/resources'
+import { SERVICE_NAME } from '.'
 import { SemanticResourceAttributes } from '@opentelemetry/semantic-conventions'
 import { ZoneContextManager } from '@opentelemetry/context-zone'
 import { getWebAutoInstrumentations } from '@opentelemetry/auto-instrumentations-web'
 import { instrumentGlobalFetch } from '$lib/util/fetch-proxy';
-import { isRedirect } from '$lib/util/types';
 import { registerInstrumentations } from '@opentelemetry/instrumentation'
 
 export * from '.';
-
-
-/**
- * Very minimal instrumentation here, because the auto-instrumentation handles the core stuff,
- * we just want to make sure that our trace-ID gets used and that we stamp errors with it.
- */
-export function traceFetch(fetch: () => ReturnType<Fetch>): ReturnType<Fetch> {
-  return tracer().startActiveSpan('fetch', async (span) => {
-    try {
-      return await fetch();
-    } catch (error) {
-      if (!isRedirect(error)) {
-        ensureErrorIsTraced(error, { span }, { ['app.error.source']: 'client-fetch-error' });
-      }
-      throw error;
-    } finally {
-      span.end();
-    }
-  });
-}
 
 instrumentGlobalFetch(() => {
   registerInstrumentations({
@@ -60,7 +39,7 @@ provider.addSpanProcessor(
     scheduledDelayMillis: 1000,
     maxQueueSize: 5000, // default: 2048
   }),
-)
+);
 
 // Debugging:
 // diag.setLogger(new DiagConsoleLogger(), DiagLogLevel.DEBUG)
@@ -73,4 +52,4 @@ provider.register({
   // Apparently shouldn't work due to zone.js if targeting ES2017+: https://github.com/open-telemetry/opentelemetry-js/tree/main/packages/opentelemetry-context-zone-peer-dep#installation
   // E.g.: https://github.com/open-telemetry/opentelemetry-js/issues/3171
   contextManager: new ZoneContextManager(),
-})
+});
